+++ nselib/capital_market/__init__.py
@@ -1,7 +1,9 @@
 from .capital_market_data import price_volume_and_deliverable_position_data, price_volume_data, \
     deliverable_position_data, bulk_deal_data, block_deals_data, short_selling_data, \
-    bhav_copy_with_delivery, bhav_copy_equities, equity_list, fno_equity_list, nifty50_equity_list, niftynext50_equity_list, niftymidcap150_equity_list, niftysmallcap250_equity_list, india_vix_data, \
-    index_data, market_watch_all_indices
+    bhav_copy_with_delivery, bhav_copy_equities, equity_list, fno_equity_list, fno_index_list, nifty50_equity_list, \
+    niftynext50_equity_list, niftymidcap150_equity_list, niftysmallcap250_equity_list, india_vix_data, \
+    index_data, market_watch_all_indices, var_begin_day, var_1st_intra_day, var_2nd_intra_day, var_3rd_intra_day, \
+    var_4th_intra_day, var_end_of_day, sme_bhav_copy, sme_band_complete, week_52_high_low_report