+++ nselib/derivatives/derivative_data.py
@@ -43,15 +43,19 @@ def future_price_volume_data(symbol: str, instrument: str, from_date: str = None
                                                from_date=start_date, to_date=end_date)
         from_date = from_date + dt.timedelta(91)
         load_days = (to_date - from_date).days
-        nse_df = pd.concat([nse_df, data_df], ignore_index=True)
+        if nse_df.empty:
+            nse_df = data_df
+        else:
+            nse_df = pd.concat([nse_df, data_df], ignore_index=True)
     return nse_df
 
 
 def get_future_price_volume_data(symbol: str, instrument: str, from_date: str, to_date: str):
+    origin_url = "https://www.nseindia.com/report-detail/fo_eq_security"
     url = "https://www.nseindia.com/api/historical/foCPV?"
     payload = f"from={from_date}&to={to_date}&instrumentType={instrument}&symbol={symbol}&csv=true"
     try:
-        data_dict = nse_urlfetch(url + payload).json()
+        data_dict = nse_urlfetch(url + payload, origin_url=origin_url).json()
     except Exception as e:
         raise ValueError(f" Invalid parameters : NSE error:{e}")
     data_df = pd.DataFrame(data_dict['data']).drop(columns='TIMESTAMP')
@@ -100,7 +104,10 @@ def option_price_volume_data(symbol: str, instrument: str, option_type: str = No
         for opt_typ in option_type:
             data_df = get_option_price_volume_data(symbol=symbol, instrument=instrument, option_type=opt_typ,
                                                    from_date=start_date, to_date=end_date)
-            nse_df = pd.concat([nse_df, data_df], ignore_index=True)
+            if nse_df.empty:
+                nse_df = data_df
+            else:
+                nse_df = pd.concat([nse_df, data_df], ignore_index=True)
         from_date = from_date + dt.timedelta(91)
         load_days = (to_date - from_date).days
 
@@ -108,11 +115,12 @@ def option_price_volume_data(symbol: str, instrument: str, option_type: str = No
 
 
 def get_option_price_volume_data(symbol: str, instrument: str, option_type: str, from_date: str, to_date: str):
+    origin_url = "https://www.nseindia.com/report-detail/fo_eq_security"
     url = "https://www.nseindia.com/api/historical/foCPV?"
     payload = f"from={from_date}&to={to_date}&instrumentType={instrument}&symbol={symbol}" \
               f"&optionType={option_type}&csv=true"
     try:
-        data_dict = nse_urlfetch(url + payload).json()
+        data_dict = nse_urlfetch(url + payload, origin_url=origin_url).json()
     except Exception as e:
         raise ValueError(f" Invalid parameters : NSE error : {e}")
     data_df = pd.DataFrame(data_dict['data']).drop(columns='TIMESTAMP')
@@ -123,14 +131,13 @@ def get_option_price_volume_data(symbol: str, instrument: str, option_type: str,
 
 def fno_bhav_copy(trade_date: str):
     """
-    NSE future option bhav copy from 2008 on wards
-    :param trade_date: eg:'01-06-2023'
+    new CM-UDiFF Common NSE future option bhav copy from 2018 on wards
+    :param trade_date: eg:'20-06-2023'
     :return: pandas Data frame
     """
     trade_date = datetime.strptime(trade_date, dd_mm_yyyy)
-    url = 'https://archives.nseindia.com/content/historical/DERIVATIVES/'
-    payload = f"{str(trade_date.strftime('%Y'))}/{str(trade_date.strftime('%b').upper())}/" \
-              f"fo{str(trade_date.strftime('%d%b%Y').upper())}bhav.csv.zip"
+    url = 'https://nsearchives.nseindia.com/content/fo/BhavCopy_NSE_FO_0_0_0_'
+    payload = f"{str(trade_date.strftime('%Y%m%d'))}_F_0000.csv.zip"
     request_bhav = nse_urlfetch(url + payload)
     bhav_df = pd.DataFrame()
     if request_bhav.status_code == 200:
@@ -139,7 +146,7 @@ def fno_bhav_copy(trade_date: str):
             if file_name:
                 bhav_df = pd.read_csv(zip_bhav.open(file_name))
     elif request_bhav.status_code == 403:
-        url2="https://www.nseindia.com/api/reports?archives=" \
+        url2 = "https://www.nseindia.com/api/reports?archives=" \
              "%5B%7B%22name%22%3A%22F%26O%20-%20Bhavcopy(csv)%22%2C%22type%22%3A%22archives%22%2C%22category%22" \
              f"%3A%22derivatives%22%2C%22section%22%3A%22equity%22%7D%5D&date={str(trade_date.strftime('%d-%b-%Y'))}" \
              f"&type=equity&mode=single"
@@ -151,28 +158,31 @@ def fno_bhav_copy(trade_date: str):
                     bhav_df = pd.read_csv(zip_bhav.open(file_name))
         elif request_bhav.status_code == 403:
             raise FileNotFoundError(f' Data not found, change the date...')
-    bhav_df = bhav_df[['INSTRUMENT', 'SYMBOL', 'EXPIRY_DT', 'STRIKE_PR', 'OPTION_TYP', 'OPEN', 'HIGH', 'LOW',
-                       'CLOSE', 'SETTLE_PR', 'CONTRACTS', 'VAL_INLAKH', 'OPEN_INT', 'CHG_IN_OI', 'TIMESTAMP']]
+    # bhav_df = bhav_df[['INSTRUMENT', 'SYMBOL', 'EXPIRY_DT', 'STRIKE_PR', 'OPTION_TYP', 'OPEN', 'HIGH', 'LOW',
+    #                    'CLOSE', 'SETTLE_PR', 'CONTRACTS', 'VAL_INLAKH', 'OPEN_INT', 'CHG_IN_OI', 'TIMESTAMP']]
     return bhav_df
 
 
 def participant_wise_open_interest(trade_date: str):
     """
     get FII, DII, Pro, Client wise participant OI data as per traded date
-    :param trade_date: eg:'01-06-2023'
+    :param trade_date: eg:'20-06-2023'
     :return: pandas Data frame
     """
     trade_date = datetime.strptime(trade_date, dd_mm_yyyy)
-    url = f"https://archives.nseindia.com/content/nsccl/fao_participant_oi_{str(trade_date.strftime('%d%m%Y'))}.csv"
+    url = f"https://nsearchives.nseindia.com/content/nsccl/fao_participant_oi_{str(trade_date.strftime('%d%m%Y'))}.csv"
     # payload = f"{str(for_date.strftime('%d%m%Y'))}.csv"
-    file_chk = requests.get(url, headers=header)
+    file_chk = nse_urlfetch(url)
+    if file_chk.status_code == 404:
+        url = f"https://archives.nseindia.com/content/nsccl/fao_participant_oi_{str(trade_date.strftime('%d%m%Y'))}.csv"
+        file_chk = nse_urlfetch(url)
     if file_chk.status_code != 200:
         raise FileNotFoundError(f" No data available for : {trade_date}")
     try:
-        data_df = pd.read_csv(url, engine='python', sep=',', quotechar='"', on_bad_lines='skip', skiprows=1,
-                              skipfooter=1)
+        # data_df = pd.read_csv(url, engine='python', sep=',', quotechar='"', on_bad_lines='skip', skiprows=1)
+        data_df = pd.read_csv(BytesIO(file_chk.content), on_bad_lines='skip', skiprows=1)
     except:
-        data_df = pd.read_csv(url, engine='c', sep=',', quotechar='"', on_bad_lines='skip', skiprows=1)
+        data_df = pd.read_csv(BytesIO(file_chk.content), on_bad_lines='skip', skiprows=1)
         data_df.drop(data_df.tail(1).index, inplace=True)
         data_df.columns = [name.replace('\t', '') for name in data_df.columns]
     return data_df
@@ -181,20 +191,20 @@ def participant_wise_open_interest(trade_date: str):
 def participant_wise_trading_volume(trade_date: str):
     """
     get FII, DII, Pro, Client wise participant volume data as per traded date
-    :param trade_date: eg:'01-06-2023'
+    :param trade_date: eg:'20-06-2023'
     :return: pandas Data frame
     """
     trade_date = datetime.strptime(trade_date, dd_mm_yyyy)
-    url = f"https://archives.nseindia.com/content/nsccl/fao_participant_vol_{str(trade_date.strftime('%d%m%Y'))}.csv"
+    url = f"https://nsearchives.nseindia.com/content/nsccl/fao_participant_vol_{str(trade_date.strftime('%d%m%Y'))}.csv"
     # payload = f"{str(for_date.strftime('%d%m%Y'))}.csv"
-    file_chk = requests.get(url, headers=header)
+    file_chk = nse_urlfetch(url)
     if file_chk.status_code != 200:
         raise FileNotFoundError(f" No data available for : {trade_date}")
     try:
-        data_df = pd.read_csv(url, engine='python', sep=',', quotechar='"', on_bad_lines='skip', skiprows=1,
-                              skipfooter=1)
+        data_df = pd.read_csv(BytesIO(file_chk.content), on_bad_lines='skip', skiprows=1)
     except Exception:
-        data_df = pd.read_csv(url, engine='c', sep=',', quotechar='"', on_bad_lines='skip', skiprows=1)
+        data_df = pd.read_csv(BytesIO(file_chk.content), engine='c', sep=',', quotechar='"',
+                              on_bad_lines='skip', skiprows=1)
         data_df.drop(data_df.tail(1).index, inplace=True)
         data_df.columns = [name.replace('\t', '') for name in data_df.columns]
     return data_df
@@ -203,14 +213,17 @@ def participant_wise_trading_volume(trade_date: str):
 def fii_derivatives_statistics(trade_date: str):
     """
     get FII derivatives statistics as per the traded date provided
-    :param trade_date: eg:'01-06-2023'
+    :param trade_date: eg:'20-06-2023'
     :return: pandas dataframe
     """
     t_date = pd.to_datetime(trade_date, format='%d-%m-%Y')
     trade_date = t_date.strftime('%d-%b-%Y')
-    url = f"https://archives.nseindia.com/content/fo/fii_stats_{trade_date}.xls"
+    url = f"https://nsearchives.nseindia.com/content/fo/fii_stats_{trade_date}.xls"
+    file_chk = nse_urlfetch(url)
+    if file_chk.status_code != 200:
+        raise FileNotFoundError(f" No data available for : {trade_date}")
     try:
-        bhav_df = pd.read_excel(url, skiprows=3, skipfooter=10).dropna()
+        bhav_df = pd.read_excel(BytesIO(file_chk.content), skiprows=3, skipfooter=10).dropna()
         bhav_df.columns = ['fii_derivatives', 'buy_contracts', 'buy_value_in_Cr', 'sell_contracts', 'sell_value_in_Cr',
                            'open_contracts', 'open_contracts_value_in_Cr']
     except Exception as e:
@@ -219,11 +232,19 @@ def fii_derivatives_statistics(trade_date: str):
 
 
 def get_nse_option_chain(symbol):
+    """
+    get NSE option chain for the symbol
+    :param symbol: eg:'TCS'/'BANKNIFTY'
+    :return: pandas dataframe
+    """
     symbol = cleaning_nse_symbol(symbol)
+    origin_url = "https://www.nseindia.com/option-chain"
     if any(x in symbol for x in indices_list):
-        payload = nse_urlfetch('https://www.nseindia.com/api/option-chain-indices?symbol=' + symbol)
+        payload = nse_urlfetch('https://www.nseindia.com/api/option-chain-indices?symbol=' + symbol,
+                               origin_url=origin_url)
     else:
-        payload = nse_urlfetch('https://www.nseindia.com/api/option-chain-equities?symbol=' + symbol)
+        payload = nse_urlfetch('https://www.nseindia.com/api/option-chain-equities?symbol=' + symbol,
+                               origin_url=origin_url)
     return payload
 
 
@@ -249,11 +270,11 @@ def expiry_dates_option_index():
     return data_dict
 
 
-def nse_live_option_chain(symbol:str, expiry_date:str = None, oi_mode: str = "full"):
+def nse_live_option_chain(symbol: str, expiry_date: str = None, oi_mode: str = "full"):
     """
     get live nse option chain.
     :param symbol: eg:SBIN/BANKNIFTY
-    :param expiry_date: '01-06-2023'
+    :param expiry_date: '20-06-2023'
     :param oi_mode: eg: full/compact
     :return: pands dataframe
     """
@@ -327,15 +348,23 @@ def nse_live_option_chain(symbol:str, expiry_date:str = None, oi_mode: str = "fu
 
             # if oi_mode == 'full':
             #     oi_row['CALLS_Chart'], oi_row['PUTS_Chart'] = 0, 0
-            oi_data = pd.concat([oi_data, pd.DataFrame([oi_row])], ignore_index=True)
+            if oi_data.empty:
+                oi_data = pd.DataFrame([oi_row]).copy()
+            else:
+                oi_data = pd.concat([oi_data, pd.DataFrame([oi_row])], ignore_index=True)
             oi_data['Symbol'] = symbol
             oi_data['Fetch_Time'] = payload['records']['timestamp']
     return oi_data
 
 
 # if __name__ == '__main__':
-    # df = future_price_volume_data("BANKNIFTY", "FUTIDX", from_date='17-06-2023', to_date='19-06-2023', period='1D')
-    # df = nse_live_option_chain(symbol='BANKNIFTY')
+    # df = future_price_volume_data("BANKNIFTY", "FUTIDX", from_date='17-06-2023', to_date='19-06-2023', period='1W')
+    # df = option_price_volume_data('NIFTY', 'OPTIDX', period='1D')
+    # df = get_nse_option_chain(symbol='TCS')
+    # df = fii_derivatives_statistics(trade_date='16-09-2024')
+    # df = expiry_dates_option_index()
+    # df = fno_bhav_copy('17-02-2025')
     # print(df)
+    # print(df.columns)
     # print(df[df['EXPIRY_DT']=='27-Jul-2023'])